班级模块接口字段调整 获取班级详情：查询条件class_id:uint32->uint64 获取班级列表：查询条件student_id:uint32->uint64 获取班级学员列表：查询条件class_id:uint32->uint64

+++ eams-cpp/eams-c1-Home/domain/query/class/classquery.h
@@ -32,7 +32,7 @@ class classQuery : public PageQuery
 {
 	DTO_INIT(classQuery, PageQuery);
 	// 班级id
-	API_DTO_FIELD_REQUIRE(UInt32, student_id, ZH_WORDS_GETTER("class.student.id"), true);
+	API_DTO_FIELD_REQUIRE(UInt64, student_id, ZH_WORDS_GETTER("class.student.id"), true);
 };
 /*
 class classinfoQuery : public oatpp::DTO
